+++ sejongdeveloper/src/main/java/com/codesoom/sejongdeveloper/dto/ReleaseOrderDetailSaveRequest.java
@@ -3,17 +3,15 @@ package com.codesoom.sejongdeveloper.dto;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
-
 @Getter
 public class ReleaseOrderDetailSaveRequest {
 
     private Long obtainOrderDetailId;    //수주 상세 일련번호
 
-    private BigDecimal quantity;    //출고수량
+    private Double quantity;    //출고수량
 
     @Builder
-    public ReleaseOrderDetailSaveRequest(Long obtainOrderDetailId, BigDecimal quantity) {
+    public ReleaseOrderDetailSaveRequest(Long obtainOrderDetailId, Double quantity) {
         this.obtainOrderDetailId = obtainOrderDetailId;
         this.quantity = quantity;
     }
